+++ gems/smorg-rb/lib/smorg/rb.rb
@@ -14,6 +14,7 @@ require 'kettle/jem/tasks/install_task'
 require 'markly/merge'
 require 'plain-merge'
 require 'prism/merge'
+require 'yaml/merge'
 require_relative 'rb/version'
 
 module Smorg
@@ -646,7 +647,9 @@ module Smorg
         '*.json5 merge=smorg-rb diff=smorg-rb smorg.language=json5',
         '*.md merge=smorg-rb diff=smorg-rb smorg.language=markdown',
         '*.markdown merge=smorg-rb diff=smorg-rb smorg.language=markdown',
-        '*.rb merge=smorg-rb diff=smorg-rb smorg.language=ruby'
+        '*.rb merge=smorg-rb diff=smorg-rb smorg.language=ruby',
+        '*.yml merge=smorg-rb diff=smorg-rb smorg.language=yaml',
+        '*.yaml merge=smorg-rb diff=smorg-rb smorg.language=yaml'
       ].each { |line| stdout.puts(line) }
       EXIT_SUCCESS
     end
@@ -697,6 +700,22 @@ module Smorg
             family: 'ruby',
             dialect: 'ruby',
             backend: 'prism',
+            profile_id: 'source_preserving',
+            fallback_policy: fallback_policy,
+            conflict_marker_size: conflict_marker_size
+          )
+        )
+      when 'yaml'
+        merge3_result(
+          Ast::Merge::Git.merge3(
+            base_source: ancestor_source,
+            ours_source: current_source,
+            theirs_source: other_source,
+            path_name: path_name,
+            provider_id: 'ruby.yaml',
+            family: 'yaml',
+            dialect: 'yaml',
+            backend: 'kreuzberg-language-pack',
             profile_id: 'source_preserving',
             fallback_policy: fallback_policy,
             conflict_marker_size: conflict_marker_size
@@ -909,6 +928,8 @@ module Smorg
         'markdown'
       when 'ruby', 'rb', 'application/x-ruby'
         'ruby'
+      when 'yaml', 'yml', 'application/yaml', 'text/yaml'
+        'yaml'
       when 'plain', 'text', 'plaintext', 'text/plain'
         'text'
       else